+++ Assets/Scripts/SaveData.cs
@@ -8,7 +8,7 @@ public class SaveData
     public int Day;
     public string Quality;
     public Resolution Resolution;
-    public float Happiness;
+    public int Happiness;
     public int Points;
     public float BackgroundVol;
     public float EffectsVol;
@@ -29,7 +29,7 @@ public class SaveData
     }
 
 
-    public SaveData(string name, int day, string quality, Resolution resolution, float happiness, int points, float backgroundVol, float effectsVol, float voicesVol, bool fullscreen)
+    public SaveData(string name, int day, string quality, Resolution resolution, int happiness, int points, float backgroundVol, float effectsVol, float voicesVol, bool fullscreen)
     {
         Name = name;
         Day = day;
